pyproject.toml
[project]
name = "cadence"
version = "0.1.0"
description = "AI-powered task prioritization agent built on LangChain Deep Agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Andy Browning" }]

dependencies = [
    "deepagents>=0.2",
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai",
    "langchain-anthropic",
    "python-dotenv",
    "rich>=13.0",
    "prompt_toolkit>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[project.scripts]
cadence = "src.agent:run_cadence"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
